alx_travel_app/requirements.txt
Django==5.2.3
djangorestframework==3.16.0
drf-yasg==1.21.10
django-cors-headers==4.7.0
django-environ==0.12.0

celery==5.5.3
requests

mysqlclient==2.2.7

